[new-style] added samplers to ShapeDef and designed a way to sample default shapes in principle
- In general, I don't think there is a need for default values (vs.
sampled values). A constant value is just a sampled value drawn from a
sample space with only one element.
- Therefore, I think shape def should include samplers in place of their
default values and we can sample them in a principled manner.
- The question now is how to not sasmple anything manually in Translation. If translation needs to have a sampled property value for one of the varying params, it should look for a readily genereated value from an instance of the corresponding shapedef, NOT doing the sampling by itself.
- In a long run, we would expose some syntax for the shapedef writers to
specify how to sample their values, which is not limited to only
numbers but all the available types in Style.