bug: walk-on-stars, dinoshade examples still failing
After debugging and fixing some small issues with the protocol, we found
that these two examples will compile in time, but the main thread would
not respond to the label request. Even worse, we cannot repro this on
webkit for walk on star.

One thing we fixed was the initial handshake. Now the main thread waits
for the worker to finish intiialization via `setTimeout`.