fix: race condition on startup run calls
the bug occurs when the IDE is loaded with an immediate run request to the worker. the race condition exists between the loading of the worker script and the run request message. if the message gets to the worker before the worker sets up the onmessage listener, the application will hang